move back the mok statsd client to tests/
